Split nab/_run.py so cache and config load no resolver
