+++ tests/test_cli.py
@@ -40,16 +40,14 @@ from nab._lock import (
     _emit_pylock,
     lock,
 )
-from nab._run import (
-    ConfigLadder,
-    _default_cache_dir,
+from nab._resolve import (
     _make_resolve_transport,
     _make_transport,
     _resolve,
-    read_config_ladder,
     resolve_extra_selection,
     resolve_group_selection,
 )
+from nab._run import ConfigLadder, _default_cache_dir, read_config_ladder
 from nab.cli import _system_exit_status, console_entry, main, run
 from nab.output import Printer, ProgressReporter, Verbosity
 from nab_index.atomic import atomic_write_text
@@ -679,7 +677,9 @@ class TestLockCommandSpecific:
         """Default format writes a real pylock.toml at the requested path."""
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out)
         text = out.read_text()
         assert 'lock-version = "1.0"' in text
@@ -691,7 +691,9 @@ class TestLockCommandSpecific:
         """No --output: pylock format defaults to pylock.toml."""
         monkeypatch.chdir(tmp_path)
         pyproject = _make_pyproject(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject)
         assert (tmp_path / "pylock.toml").exists()
 
@@ -701,7 +703,9 @@ class TestLockCommandSpecific:
         """No --output: requirements format defaults to requirements.txt."""
         monkeypatch.chdir(tmp_path)
         pyproject = _make_pyproject(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, format="requirements")
         text = (tmp_path / "requirements.txt").read_text()
         assert "foo==1.0" in text
@@ -713,7 +717,9 @@ class TestLockCommandSpecific:
         """requirements-without-hashes defaults to requirements.txt."""
         monkeypatch.chdir(tmp_path)
         pyproject = _make_pyproject(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, format="requirements-without-hashes")
         text = (tmp_path / "requirements.txt").read_text()
         assert "foo==1.0" in text
@@ -725,7 +731,9 @@ class TestLockCommandSpecific:
         """A build lock defaults clear of the runtime lock's name."""
         monkeypatch.chdir(tmp_path)
         pyproject = _make_pyproject(tmp_path, _BUILD_SYSTEM_PROJECT)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, build_requirements=True)
         assert (tmp_path / "pylock.build.toml").exists()
         assert not (tmp_path / "pylock.toml").exists()
@@ -739,7 +747,9 @@ class TestLockCommandSpecific:
         """Both requirements formats get their own default name."""
         monkeypatch.chdir(tmp_path)
         pyproject = _make_pyproject(tmp_path, _BUILD_SYSTEM_PROJECT)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, format=lock_format, build_requirements=True)
         assert (tmp_path / "build-requirements.txt").exists()
         assert not (tmp_path / "requirements.txt").exists()
@@ -761,7 +771,9 @@ class TestLockCommandSpecific:
             "[tool.nab]\ncreated-at = 2020-01-01T00:00:00+00:00\n"
         )
         pyproject = _make_pyproject(tmp_path, _BUILD_SYSTEM_PROJECT)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, build_requirements=True)
         written = tomli.loads((tmp_path / "pylock.build.toml").read_text())
         assert written["tool"]["nab"]["created-at"] == recorded
@@ -776,7 +788,9 @@ class TestLockCommandSpecific:
             '[tool.nab]\ndefault-groups = ["dev"]\nbase-group = "default"\n',
         )
         out = tmp_path / "pylock.build.toml"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out, build_requirements=True)
         written = tomli.loads(out.read_text())
         assert "default-groups" not in written
@@ -792,7 +806,7 @@ class TestLockCommandSpecific:
         )
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_stub_resolve_result(pins={"foo": V("1.0")}),
         ):
             lock(pyproject, output=out)
@@ -872,7 +886,9 @@ class TestLockCommandSpecific:
         """`requirements` format renders --hash lines."""
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "requirements.txt"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out, format="requirements")
         text = out.read_text()
         assert "foo==1.0" in text
@@ -941,7 +957,9 @@ class TestLockCommandSpecific:
         """requirements-without-hashes renders one name==version per line."""
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "requirements.txt"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out, format="requirements-without-hashes")
         text = out.read_text()
         assert text.strip() == "foo==1.0"
@@ -951,7 +969,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "requirements.txt"
         result = _hashless_resolve_result()
-        with patch("nab._run.resolve_for_targets", return_value=result):
+        with patch("nab._resolve.resolve_for_targets", return_value=result):
             lock(pyproject, output=out, format="requirements-without-hashes")
         assert out.read_text().strip() == "foo==1.0"
 
@@ -963,7 +981,7 @@ class TestLockCommandSpecific:
         out = tmp_path / "pylock.toml"
         result = _hashless_resolve_result()
         with (
-            patch("nab._run.resolve_for_targets", return_value=result),
+            patch("nab._resolve.resolve_for_targets", return_value=result),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=out, format="pylock")
@@ -980,7 +998,7 @@ class TestLockCommandSpecific:
             target_results=[_failed(target, ResolutionError("conflict"))],
         )
         with (
-            patch("nab._run.resolve_for_targets", return_value=result),
+            patch("nab._resolve.resolve_for_targets", return_value=result),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=tmp_path / "pylock.toml")
@@ -1016,7 +1034,7 @@ class TestLockCommandSpecific:
             ],
         )
         with (
-            patch("nab._run.resolve_for_targets", return_value=result),
+            patch("nab._resolve.resolve_for_targets", return_value=result),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(
@@ -1035,7 +1053,9 @@ class TestLockCommandSpecific:
     ) -> None:
         """`--output -` routes pylock format to stdout."""
         pyproject = _make_pyproject(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=Path("-"))
         out = capsys.readouterr().out
         assert 'lock-version = "1.0"' in out
@@ -1046,7 +1066,9 @@ class TestLockCommandSpecific:
     ) -> None:
         """`--output -` routes requirements format to stdout."""
         pyproject = _make_pyproject(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=Path("-"), format="requirements")
         out = capsys.readouterr().out
         assert "foo==1.0" in out
@@ -1057,7 +1079,9 @@ class TestLockCommandSpecific:
     ) -> None:
         """`--output -` routes requirements-without-hashes to stdout."""
         pyproject = _make_pyproject(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=Path("-"), format="requirements-without-hashes")
         assert capsys.readouterr().out.strip() == "foo==1.0"
 
@@ -1067,7 +1091,8 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets", side_effect=ResolutionError("conflict")
+                "nab._resolve.resolve_for_targets",
+                side_effect=ResolutionError("conflict"),
             ),
             pytest.raises(SystemExit, match="1"),
         ):
@@ -1094,7 +1119,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=UnsupportedVcsError("refusing direct-URL requirement"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -1109,7 +1134,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=InvalidUploadTimeError("foo 1.0 has a naive upload time"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -1124,7 +1149,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=NotImplementedError("resolver path is not implemented"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -1139,7 +1164,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=ConfigError("Constraints cannot have extras: idna[foo]<3"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -1152,7 +1177,9 @@ class TestLockCommandSpecific:
     ) -> None:
         pyproject = _make_pyproject(tmp_path)
         with (
-            patch("nab._run.resolve_for_targets", side_effect=KeyError("dependencies")),
+            patch(
+                "nab._resolve.resolve_for_targets", side_effect=KeyError("dependencies")
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject)
@@ -1186,7 +1213,8 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets", side_effect=MissingHashError("no hash")
+                "nab._resolve.resolve_for_targets",
+                side_effect=MissingHashError("no hash"),
             ),
             pytest.raises(SystemExit, match="1"),
         ):
@@ -1200,7 +1228,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=InvalidProjectRequirementError("invalid requirement 'x y'"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -1215,7 +1243,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=HttpError("GET https://pypi.org/simple/foo/ failed: 503"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -1240,7 +1268,7 @@ class TestLockCommandSpecific:
 
         pyproject = _make_pyproject(tmp_path)
         with (
-            patch("nab._run.resolve_for_targets", side_effect=caught.value),
+            patch("nab._resolve.resolve_for_targets", side_effect=caught.value),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject)
@@ -1281,7 +1309,7 @@ class TestLockCommandSpecific:
 
         pyproject = _make_pyproject(tmp_path)
         with (
-            patch("nab._run.resolve_for_targets", side_effect=caught.value),
+            patch("nab._resolve.resolve_for_targets", side_effect=caught.value),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject)
@@ -1297,7 +1325,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=MissingSdistError("foo==1.0 has no sdist"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -1316,7 +1344,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=LookupError("unknown group 'ghost'"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -1331,7 +1359,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=MissingExtraError(
                     "foo==1.0 does not provide extra 'nonexistent'"
                 ),
@@ -1404,7 +1432,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=SiblingMetadataDivergenceError(
                     "foo 1.0 has tie-ranked wheels foo-1.0-cp311.cp312-none-any.whl "
                     "and foo-1.0-cp312.cp313-none-any.whl that declare different "
@@ -1456,7 +1484,7 @@ class TestLockCommandSpecific:
         )
 
         with (
-            patch("nab._run._make_transport", return_value=transport),
+            patch("nab._resolve._make_transport", return_value=transport),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=tmp_path / "pylock.toml", cache=False)
@@ -1519,7 +1547,7 @@ class TestLockCommandSpecific:
         transport = _SidecarTransport(bodies)
 
         with (
-            patch("nab._run._make_transport", return_value=transport),
+            patch("nab._resolve._make_transport", return_value=transport),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=tmp_path / "pylock.toml", cache=False)
@@ -1566,7 +1594,7 @@ class TestLockCommandSpecific:
         )
 
         with (
-            patch("nab._run._make_transport", return_value=transport),
+            patch("nab._resolve._make_transport", return_value=transport),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=tmp_path / "pylock.toml", cache=False)
@@ -1612,7 +1640,7 @@ class TestLockCommandSpecific:
         )
 
         with (
-            patch("nab._run._make_transport", return_value=transport),
+            patch("nab._resolve._make_transport", return_value=transport),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=tmp_path / "pylock.toml", cache=False)
@@ -1915,7 +1943,9 @@ class TestLockCommandSpecific:
         out = tmp_path / "pylock.toml"
         out.mkdir()
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=out)
@@ -1928,7 +1958,9 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "nope" / "pylock.toml"
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=out)
@@ -1946,7 +1978,9 @@ class TestLockCommandSpecific:
         committed = b'lock-version = "1.0"\n'
         out.write_bytes(committed)
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             cap_writes(64),
             pytest.raises(SystemExit, match="1"),
         ):
@@ -1959,7 +1993,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
         ) as mock_resolve:
             lock(pyproject, output=out, project_resolution="lowest")
         kwargs = mock_resolve.call_args.kwargs
@@ -1970,7 +2004,7 @@ class TestLockCommandSpecific:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
         ) as mock_resolve:
             lock(pyproject, output=out)
         assert mock_resolve.call_args.kwargs["resolution_strategy"] is None
@@ -1988,7 +2022,9 @@ class TestLockCommandSpecific:
             "nab._run._config_search_roots",
             lambda p: SourceRoots(project_dir=p.parent, pyproject=p),
         )
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out, project_resolution="lowest")
         err = capsys.readouterr().err
         assert "does not derive from the committed" in err
@@ -2007,7 +2043,9 @@ class TestLockCommandSpecific:
             "nab._run._config_search_roots",
             lambda p: SourceRoots(project_dir=p.parent, pyproject=p),
         )
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out)
         assert "does not derive from the committed" not in capsys.readouterr().err
 
@@ -2029,7 +2067,9 @@ class TestLockCommandSpecific:
             lambda p: SourceRoots(project_dir=p.parent, pyproject=p),
         )
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=out)
@@ -2058,7 +2098,9 @@ class TestLockCommandSpecific:
             lambda p: SourceRoots(user_toml=user, project_dir=p.parent, pyproject=p),
         )
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=out)
@@ -2081,7 +2123,9 @@ class TestLockCommandSpecific:
             lambda p: SourceRoots(user_toml=user, project_dir=p.parent, pyproject=p),
         )
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, output=out)
@@ -2097,7 +2141,7 @@ class TestPythonFlag:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
         ) as mock_resolve:
             lock(pyproject, output=out, python="3.11")
         config = mock_resolve.call_args.kwargs["config"]
@@ -2107,7 +2151,7 @@ class TestPythonFlag:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
         ) as mock_resolve:
             lock(pyproject, output=out)
         assert mock_resolve.call_args.kwargs["config"].environment is None
@@ -2157,7 +2201,7 @@ class TestPythonFlag:
         out = tmp_path / "wheels"
         with (
             patch(
-                "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
             ) as mock_resolve,
             patch(
                 "nab._download.download_lock",
@@ -2188,7 +2232,7 @@ class TestPythonFlag:
         )
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
         ) as mock_resolve:
             lock(pyproject, output=out, python="3.14")
         assert mock_resolve.call_args.kwargs["config"].environment.python == "3.14"
@@ -2298,7 +2342,7 @@ class TestProjectFlagErrors:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
         ) as mock_resolve:
             lock(pyproject, output=out, project_requires_python="==3.11")
         assert mock_resolve.call_args.kwargs["config"].requires_python == "==3.11"
@@ -2329,7 +2373,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "pylock.toml"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=InvalidProjectRequirementError("invalid requirement 'x y'"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2345,7 +2389,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "pylock.toml"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=InvalidUploadTimeError("foo 1.0 has a naive upload time"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2361,7 +2405,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "pylock.toml"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=ConfigError(
                     "[tool.nab].conflicts names extra 'gpuu', which the project"
                     " does not declare in [project.optional-dependencies]"
@@ -2382,7 +2426,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "pylock.toml"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=NotImplementedError("resolver path is not implemented"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2486,7 +2530,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ):
             lock(pyproject, output=out)
@@ -2507,7 +2551,7 @@ class TestLockCommandUniversal:
         out.write_bytes(committed)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             cap_writes(4),
@@ -2532,7 +2576,7 @@ class TestLockCommandUniversal:
         )
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ):
             lock(pyproject, output=out, groups=("test",))
@@ -2545,7 +2589,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ) as mock_resolve:
             lock(pyproject, output=out, http_backend="urllib3", offline=True)
@@ -2560,7 +2604,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "requirements.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ):
             lock(pyproject, format="requirements", output=out)
@@ -2576,7 +2620,7 @@ class TestLockCommandUniversal:
         """Universal + pylock + --output - writes lock text to stdout."""
         pyproject = _universal_pyproject(tmp_path)
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ):
             lock(pyproject, output=Path("-"))
@@ -2590,7 +2634,7 @@ class TestLockCommandUniversal:
         monkeypatch.chdir(tmp_path)
         pyproject = _universal_pyproject(tmp_path)
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ):
             lock(pyproject)
@@ -2603,7 +2647,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             patch("nab._lock.write_lock", side_effect=MissingHashError("no hash")),
@@ -2628,7 +2672,7 @@ class TestLockCommandUniversal:
         )
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             patch("nab._lock.write_lock", side_effect=DisjointnessError(hint)),
@@ -2646,7 +2690,7 @@ class TestLockCommandUniversal:
         message = "conflict-respecting selections exceed 100000"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             patch("nab._lock.write_lock", side_effect=IntractableMarkerError(message)),
@@ -2729,7 +2773,7 @@ class TestLockCommandUniversal:
         )
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             patch(
@@ -2773,7 +2817,7 @@ class TestLockCommandUniversal:
         )
 
         with (
-            patch("nab._run.resolve_for_targets", return_value=result),
+            patch("nab._resolve.resolve_for_targets", return_value=result),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(
@@ -2793,7 +2837,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=UnsupportedVcsError("refusing direct-URL requirement"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2813,7 +2857,7 @@ class TestLockCommandUniversal:
         """
         pyproject = _universal_pyproject(tmp_path)
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_multi_tuple_universal_result(),
         ):
             lock(pyproject, format="requirements-without-hashes")
@@ -2835,7 +2879,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "pins.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ):
             lock(pyproject, format="requirements-without-hashes", output=out)
@@ -2849,7 +2893,7 @@ class TestLockCommandUniversal:
         """An explicit ``--output -`` prints the same per-tuple blocks."""
         pyproject = _universal_pyproject(tmp_path)
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_multi_tuple_universal_result(),
         ):
             lock(pyproject, format="requirements-without-hashes", output=Path("-"))
@@ -2864,7 +2908,7 @@ class TestLockCommandUniversal:
         """Single-tuple matrix + ``--output -``: just the pins, no header."""
         pyproject = _universal_pyproject(tmp_path)
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ):
             lock(pyproject, format="requirements-without-hashes", output=Path("-"))
@@ -2877,7 +2921,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_multi_tuple_failed_result(ResolutionError("conflict")),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2895,7 +2939,7 @@ class TestLockCommandUniversal:
         multi = "first line\nsecond line\nthird line"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_multi_tuple_failed_result(ResolutionError(multi)),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2913,7 +2957,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_multi_tuple_failed_result(None),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2933,7 +2977,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(
                     success=False, error=ResolutionError("conflict")
                 ),
@@ -2952,7 +2996,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=KeyError("dependencies"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2967,7 +3011,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=LookupError("unknown group 'ghost'"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2982,7 +3026,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=HttpError("GET https://pypi.org/simple/foo/ failed: 503"),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -2999,7 +3043,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=MissingExtraError(
                     "foo==1.0 does not provide extra 'nonexistent'"
                 ),
@@ -3022,7 +3066,7 @@ class TestLockCommandUniversal:
 
         pyproject = _universal_pyproject(tmp_path)
         with (
-            patch("nab._run.resolve_for_targets", side_effect=caught.value),
+            patch("nab._resolve.resolve_for_targets", side_effect=caught.value),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, format="requirements-without-hashes")
@@ -3037,7 +3081,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             patch(
@@ -3069,7 +3113,7 @@ class TestLockCommandUniversal:
             ],
         )
         with (
-            patch("nab._run.resolve_for_targets", return_value=mixed),
+            patch("nab._resolve.resolve_for_targets", return_value=mixed),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, format="requirements-without-hashes")
@@ -3102,7 +3146,7 @@ class TestLockCommandUniversal:
             ],
         )
         with (
-            patch("nab._run.resolve_for_targets", return_value=mixed),
+            patch("nab._resolve.resolve_for_targets", return_value=mixed),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, format="requirements-without-hashes")
@@ -3233,7 +3277,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "constraints-{python_version}.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_multi_tuple_universal_result(),
         ):
             lock(pyproject, format="requirements-without-hashes", output=out)
@@ -3253,7 +3297,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "constraints-{python_version}-{platform_id}.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_multi_tuple_universal_result(),
         ):
             lock(pyproject, format="requirements-without-hashes", output=out)
@@ -3267,7 +3311,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "req-{python_version}.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_multi_tuple_universal_result(),
         ):
             lock(pyproject, format="requirements", output=out)
@@ -3293,7 +3337,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "requirements.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_multi_tuple_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3321,7 +3365,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "constraints-{python_version}.txt"
         with (
-            patch("nab._run.resolve_for_targets", return_value=result),
+            patch("nab._resolve.resolve_for_targets", return_value=result),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, format="requirements-without-hashes", output=out)
@@ -3338,7 +3382,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "req-{python_version}-{foo}.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3356,7 +3400,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "req-{python_version}-{.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3372,7 +3416,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "req-{python_version}-{platform_id:d}.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3389,7 +3433,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "req-{platform_id}-{python_version!r}.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3403,7 +3447,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "constraints-{python_version}.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_universal_result(success=True),
         ):
             lock(pyproject, format="requirements-without-hashes", output=out)
@@ -3416,7 +3460,7 @@ class TestLockCommandUniversal:
         pyproject = _universal_pyproject(tmp_path)
         out = tmp_path / "req-{selection}.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_forked_universal_result(),
         ):
             lock(pyproject, format="requirements-without-hashes", output=out)
@@ -3438,7 +3482,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "req-{python_version}-{platform_id}.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_forked_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3457,7 +3501,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "requirements.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_forked_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3481,7 +3525,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "req-{python_version}-{platform_id}-{selection}.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_two_libc_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3500,7 +3544,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "requirements.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_two_libc_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3518,7 +3562,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "requirements.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_two_implementation_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3539,7 +3583,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "req-{platform_id}.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_mixed_implementation_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3560,7 +3604,7 @@ class TestLockCommandUniversal:
         out = tmp_path / "req-{python_version}.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             patch(
@@ -3586,7 +3630,7 @@ class TestLockCommandUniversal:
         first.write_text("stale==0.1\n")
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_late_hashless_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3613,7 +3657,7 @@ class TestLockCommandUniversal:
         (tmp_path / "req-3.12.txt").mkdir()
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_multi_tuple_universal_result(),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3635,7 +3679,7 @@ class TestLockCommandUniversal:
         first.write_text("stale==0.1\n")
         before = stat.S_IMODE(first.stat().st_mode)
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_multi_tuple_universal_result(),
         ):
             lock(pyproject, format="requirements-without-hashes", output=out)
@@ -3659,7 +3703,7 @@ class TestLockCommandUniversal:
         )
         out = tmp_path / "constraints-{python_version}.txt"
         with (
-            patch("nab._run.resolve_for_targets", return_value=mixed),
+            patch("nab._resolve.resolve_for_targets", return_value=mixed),
             pytest.raises(SystemExit, match="1"),
         ):
             lock(pyproject, format="requirements-without-hashes", output=out)
@@ -3691,7 +3735,8 @@ class TestNoEmitWorkspace:
         pyproject = _workspace_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=self._alpha_and_foo_result()
+            "nab._resolve.resolve_for_targets",
+            return_value=self._alpha_and_foo_result(),
         ):
             lock(pyproject, output=out, no_emit_workspace=True)
         text = out.read_text()
@@ -3703,7 +3748,8 @@ class TestNoEmitWorkspace:
         pyproject = _workspace_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=self._alpha_and_foo_result()
+            "nab._resolve.resolve_for_targets",
+            return_value=self._alpha_and_foo_result(),
         ):
             lock(pyproject, output=out)
         text = out.read_text()
@@ -3716,7 +3762,8 @@ class TestNoEmitWorkspace:
         pyproject = _workspace_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=self._alpha_and_foo_result()
+            "nab._resolve.resolve_for_targets",
+            return_value=self._alpha_and_foo_result(),
         ):
             lock(pyproject, output=out, no_emit_workspace=True)
         assert "(1 packages)" in capsys.readouterr().err
@@ -3726,7 +3773,8 @@ class TestNoEmitWorkspace:
         pyproject = _workspace_pyproject(tmp_path)
         out = tmp_path / "requirements.txt"
         with patch(
-            "nab._run.resolve_for_targets", return_value=self._alpha_and_foo_result()
+            "nab._resolve.resolve_for_targets",
+            return_value=self._alpha_and_foo_result(),
         ):
             lock(pyproject, output=out, format="requirements", no_emit_workspace=True)
         text = out.read_text()
@@ -3738,7 +3786,7 @@ class TestNoEmitWorkspace:
         pyproject = _workspace_pyproject(tmp_path, universal=True)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=self._alpha_and_foo_universal(),
         ):
             lock(pyproject, output=out, no_emit_workspace=True)
@@ -3752,7 +3800,7 @@ class TestNoEmitWorkspace:
         """Universal requirements + stdout drops workspace pin lines."""
         pyproject = _workspace_pyproject(tmp_path, universal=True)
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=self._alpha_and_foo_universal(),
         ):
             lock(
@@ -3772,7 +3820,7 @@ class TestNoEmitWorkspace:
         pyproject = _workspace_pyproject(tmp_path, universal=True)
         out = tmp_path / "constraints-{python_version}.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=self._alpha_and_foo_universal(),
         ):
             lock(
@@ -3792,7 +3840,7 @@ class TestNoEmitWorkspace:
         pyproject = _workspace_pyproject(tmp_path, universal=True)
         out = tmp_path / "requirements.txt"
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=self._alpha_and_foo_universal(),
         ):
             lock(
@@ -3811,7 +3859,8 @@ class TestNoEmitWorkspace:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with patch(
-            "nab._run.resolve_for_targets", return_value=self._alpha_and_foo_result()
+            "nab._resolve.resolve_for_targets",
+            return_value=self._alpha_and_foo_result(),
         ):
             lock(pyproject, output=out, no_emit_workspace=True)
         text = out.read_text()
@@ -3837,7 +3886,7 @@ class TestNoEmitWorkspace:
         )
         pyproject = _workspace_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
-        with patch("nab._run.resolve_for_targets", return_value=result):
+        with patch("nab._resolve.resolve_for_targets", return_value=result):
             lock(pyproject, output=out, no_emit_workspace=True)
         text = out.read_text()
         assert 'name = "foo"' in text
@@ -3982,7 +4031,7 @@ class TestPylockOutputNameValidation:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_universal_result(success=True),
             ),
             pytest.raises(SystemExit, match="1"),
@@ -3993,7 +4042,9 @@ class TestPylockOutputNameValidation:
     def test_specific_accepts_named_pylock(self, tmp_path: Path) -> None:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.dev.toml"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out)
         assert out.exists()
 
@@ -4034,7 +4085,9 @@ class TestPylockOutputNameValidation:
         self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
     ) -> None:
         pyproject = _make_pyproject(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=Path("-"))
         assert 'lock-version = "1.0"' in capsys.readouterr().out
 
@@ -4042,7 +4095,9 @@ class TestPylockOutputNameValidation:
         """A non-pylock format is free to use any output name."""
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "constraints.txt"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out, format="requirements")
         assert out.exists()
 
@@ -4316,7 +4371,7 @@ class TestLockAnchorReuse:
         """The ``P4D`` window a re-lock over a recorded anchor resolves against."""
         prior, pyproject = self._relative_cutoff_relock(tmp_path)
         with patch(
-            "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
         ) as mock_resolve:
             lock(pyproject, output=prior, upgrade=upgrade)
 
@@ -4326,7 +4381,9 @@ class TestLockAnchorReuse:
 
     def test_relock_records_reused_anchor(self, tmp_path: Path) -> None:
         prior, pyproject = self._relative_cutoff_relock(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=prior)
         # New pylock's [tool.nab].created-at must equal the prior anchor.
         from nab_project.lockfile import read_lockfile_anchor
@@ -4335,7 +4392,9 @@ class TestLockAnchorReuse:
 
     def test_upgrade_writes_fresh_anchor(self, tmp_path: Path) -> None:
         prior, pyproject = self._relative_cutoff_relock(tmp_path)
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=prior, upgrade=True)
         from nab_project.lockfile import read_lockfile_anchor
 
@@ -4353,7 +4412,9 @@ class TestLockAnchorReuse:
             f'[tool.nab]\nuploaded-prior-to = "{absolute.isoformat()}"\n',
         )
         out = tmp_path / "pylock.toml"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out)
         from nab_project.lockfile import read_lockfile_anchor
 
@@ -4370,7 +4431,9 @@ class TestLockAnchorReuse:
             f'[tool.nab]\nuploaded-prior-to = "{absolute.isoformat()}"\n',
         )
         out = tmp_path / "pylock.toml"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             lock(pyproject, output=out)
         from nab_project.lockfile import read_lockfile_anchor
 
@@ -4469,7 +4532,7 @@ class TestLockedFlag:
         result: ResolveResult,
         *extra_args: str,
     ) -> None:
-        with patch("nab._run.resolve_for_targets", return_value=result):
+        with patch("nab._resolve.resolve_for_targets", return_value=result):
             _lock_cli(str(pyproject), "--output", str(out), *extra_args)
 
     def _run_locked(
@@ -4480,7 +4543,7 @@ class TestLockedFlag:
         *extra_args: str,
         status: int = 0,
     ) -> None:
-        with patch("nab._run.resolve_for_targets", return_value=result):
+        with patch("nab._resolve.resolve_for_targets", return_value=result):
             _lock_cli(
                 str(pyproject),
                 "--output",
@@ -4632,7 +4695,7 @@ class TestLockedFlag:
         before = out.read_bytes()
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={"foo": V("1.0")}),
             ),
             patch("nab._lock.render_lock", side_effect=MissingHashError("no hash")),
@@ -4673,7 +4736,7 @@ class TestLockedFlag:
         hint = "foo: 2 entries fire under env='py311-linux_x86_64'"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={"foo": V("1.0")}),
             ),
             patch("nab._lock.render_lock", side_effect=DisjointnessError(hint)),
@@ -4802,7 +4865,9 @@ class TestLockProvenanceCliOverrides:
     def test_cli_override_recorded_in_pylock(self, tmp_path: Path) -> None:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             _lock_cli(
                 str(pyproject), "--output", str(out), "--project-resolution", "lowest"
             )
@@ -4812,7 +4877,9 @@ class TestLockProvenanceCliOverrides:
     def test_no_cli_override_omits_key(self, tmp_path: Path) -> None:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
-        with patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()):
+        with patch(
+            "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+        ):
             _lock_cli(str(pyproject), "--output", str(out))
         block = tomli.loads(out.read_text())["tool"]["nab"]
         assert "cli-project-overrides" not in block
@@ -4822,7 +4889,9 @@ class TestLockProvenanceCliOverrides:
         out = tmp_path / "pylock.toml"
         argv = ["/somewhere/on/this/machine/src/nab/__main__.py", "lock", "--offline"]
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch.object(sys, "argv", argv),
         ):
             _lock_cli(str(pyproject), "--output", str(out))
@@ -4848,7 +4917,9 @@ class TestLockNonUtf8Text:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "pylock.toml"
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch.object(sys, "argv", self._bad_argv()),
             pytest.raises(SystemExit) as exc,
         ):
@@ -4873,7 +4944,9 @@ class TestLockNonUtf8Text:
         raw = io.BytesIO()
         stream = io.TextIOWrapper(raw, encoding="utf-8", errors="surrogateescape")
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch.object(sys, "argv", self._bad_argv()),
             patch.object(sys, "stdout", stream),
             pytest.raises(SystemExit) as exc,
@@ -4911,7 +4984,7 @@ class TestLockNonUtf8Text:
         out = tmp_path / "req-{python_version}.txt"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_multi_tuple_universal_result(),
             ),
             patch(
@@ -4933,7 +5006,9 @@ class TestLockNonUtf8Text:
         pyproject = _make_pyproject(tmp_path)
         out = tmp_path / "requirements.txt"
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch.object(sys, "argv", self._bad_argv()),
         ):
             lock(
@@ -4999,7 +5074,7 @@ class TestGroupAndExtraSelection:
         pyproject = _make_pyproject(tmp_path)
         denied = PermissionError(errno.EACCES, "Permission denied", str(pyproject))
         with (
-            patch("nab._run.read_pyproject_groups", side_effect=denied),
+            patch("nab._resolve.read_pyproject_groups", side_effect=denied),
             pytest.raises(SystemExit, match="1"),
         ):
             resolve_group_selection(pyproject, groups=(), all_groups=True)
@@ -5014,7 +5089,9 @@ class TestGroupAndExtraSelection:
         pyproject = _make_pyproject(tmp_path)
         denied = PermissionError(errno.EACCES, "Permission denied", str(pyproject))
         with (
-            patch("nab._run.read_pyproject_optional_dependencies", side_effect=denied),
+            patch(
+                "nab._resolve.read_pyproject_optional_dependencies", side_effect=denied
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             resolve_extra_selection(pyproject, extras=(), all_extras=True)
@@ -5263,7 +5340,7 @@ class TestCacheFlags:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={}),
             ) as mock_resolve,
             patch("nab._lock.write_lock"),
@@ -5279,7 +5356,7 @@ class TestCacheFlags:
         cache = tmp_path / "mycache"
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={}),
             ) as mock_resolve,
             patch("nab._lock.write_lock"),
@@ -5292,7 +5369,7 @@ class TestCacheFlags:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={}),
             ) as mock_resolve,
             patch("nab._lock.write_lock"),
@@ -5305,7 +5382,7 @@ class TestCacheFlags:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={}),
             ) as mock_resolve,
             patch("nab._lock.write_lock"),
@@ -5506,7 +5583,7 @@ class TestOfflineFlagSurface:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={}),
             ) as mock_resolve,
             patch("nab._lock.write_lock"),
@@ -5569,11 +5646,11 @@ class TestLayeredRunKnobSurface:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={}),
             ),
             patch("nab._lock.write_lock"),
-            patch("nab._run._make_transport") as mock_transport,
+            patch("nab._resolve._make_transport") as mock_transport,
         ):
             _lock_cli(
                 str(pyproject),
@@ -5681,7 +5758,7 @@ class TestMain:
         pyproject = _make_pyproject(tmp_path)
         config = read_pyproject_config(pyproject)
         with patch(
-            "nab._run.resolve_for_targets",
+            "nab._resolve.resolve_for_targets",
             return_value=_stub_resolve_result(pins={}),
         ):
             result = _resolve(
@@ -5705,7 +5782,7 @@ class TestMain:
             enabled_during.append(gc.isenabled())
             return _stub_resolve_result(pins={})
 
-        with patch("nab._run.resolve_for_targets", side_effect=record_gc_state):
+        with patch("nab._resolve.resolve_for_targets", side_effect=record_gc_state):
             _resolve(
                 pyproject,
                 config=config,
@@ -5739,7 +5816,7 @@ class TestMain:
         # generation 2 on the session's own schedule.
         gc.collect()
 
-        with patch("nab._run.resolve_for_targets", side_effect=allocate_tracked):
+        with patch("nab._resolve.resolve_for_targets", side_effect=allocate_tracked):
             _resolve(
                 pyproject,
                 config=config,
@@ -5772,7 +5849,7 @@ class TestMain:
         config = read_pyproject_config(pyproject)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=ResolutionError("no solution"),
             ),
             pytest.raises(SystemExit),
@@ -5794,7 +5871,7 @@ class TestMain:
         config = read_pyproject_config(pyproject)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=ResolutionError("no solution"),
             ),
             pytest.raises(SystemExit),
@@ -5917,7 +5994,9 @@ class TestClosedStandardStreams:
         )
         monkeypatch.setattr(sys, "stderr", None)
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch("nab.cli.os._exit") as mock_exit,
         ):
             console_entry()
@@ -6033,7 +6112,7 @@ class TestMainWiresOutputOptions:
 
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_stub_resolve_result(pins={}),
             ),
             patch("nab._lock.write_lock"),
@@ -6192,7 +6271,7 @@ class TestProgressReachesTheResolve:
         monkeypatch.setattr(sys, "stderr", stderr)
 
         with (
-            patch("nab._run.resolve_for_targets", side_effect=resolve),
+            patch("nab._resolve.resolve_for_targets", side_effect=resolve),
             command_patch,
         ):
             main()
@@ -6394,7 +6473,7 @@ class TestResolveTransport:
 
     def test_online_resolve_gets_the_transport_up_front(self) -> None:
         """A resolve that may fetch is handed the real transport."""
-        with patch("nab._run._make_transport") as make:
+        with patch("nab._resolve._make_transport") as make:
             transport = _make_resolve_transport("urllib3", offline=False)
 
         assert transport is make.return_value
@@ -6402,7 +6481,7 @@ class TestResolveTransport:
 
     def test_offline_httpx_resolve_gets_it_up_front_too(self) -> None:
         """Only urllib3 defers; httpx is built while the CLI can still exit."""
-        with patch("nab._run._make_transport") as make:
+        with patch("nab._resolve._make_transport") as make:
             transport = _make_resolve_transport("httpx", offline=True)
 
         assert transport is make.return_value
@@ -6443,7 +6522,7 @@ class TestResolveTransport:
         pyproject = _make_pyproject(
             tmp_path, '[project]\nname = "root"\nversion = "0"\ndependencies = []\n'
         )
-        with patch("nab._run._make_urllib3_transport") as make:
+        with patch("nab._resolve._make_urllib3_transport") as make:
             lock(pyproject, output=tmp_path / "pylock.toml", offline=True, cache=False)
 
         make.assert_not_called()
@@ -6451,7 +6530,7 @@ class TestResolveTransport:
     def test_first_request_builds_one_transport_and_forwards(self) -> None:
         """A deferred transport builds its inner one once, on the first get."""
         inner = _RecordingTransport()
-        with patch("nab._run._make_urllib3_transport", return_value=inner) as make:
+        with patch("nab._resolve._make_urllib3_transport", return_value=inner) as make:
             transport = _make_resolve_transport("urllib3", offline=True)
             make.assert_not_called()
 
@@ -6472,7 +6551,7 @@ class TestResolveTransport:
     def test_close_closes_the_transport_it_built(self) -> None:
         """Closing a deferred transport that fetched closes the inner one."""
         inner = _RecordingTransport()
-        with patch("nab._run._make_urllib3_transport", return_value=inner):
+        with patch("nab._resolve._make_urllib3_transport", return_value=inner):
             transport = _make_resolve_transport("urllib3", offline=True)
 
             async def fetch_then_close() -> None:
@@ -6485,7 +6564,7 @@ class TestResolveTransport:
 
     def test_close_without_a_request_builds_nothing(self) -> None:
         """Closing a deferred transport that never fetched builds nothing."""
-        with patch("nab._run._make_urllib3_transport") as make:
+        with patch("nab._resolve._make_urllib3_transport") as make:
             transport = _make_resolve_transport("urllib3", offline=True)
             asyncio.run(transport.aclose())
 
@@ -6506,6 +6585,7 @@ class TestResolveTransport:
 _PACKAGE_ENTRY_STATEMENTS = (
     "import nab.cli",
     "import nab._run",
+    "import nab._resolve",
     "import nab._config_cmd",
     "from nab._lock import lock",
     "from nab._download import download",
@@ -6517,13 +6597,11 @@ class TestPackageEntryStatements:
 
     @pytest.mark.parametrize("statement", _PACKAGE_ENTRY_STATEMENTS)
     def test_entry_statement_runs_in_a_fresh_interpreter(self, statement: str) -> None:
-        """``cli`` and each command module import each other.
+        """Each statement is the first thing its interpreter does.
 
-        A probe that enters through the command module runs ``cli``
-        against a half-built one, which a name import at either end of
-        the cycle cannot survive.  Nothing else in the suite can see
-        that: ``conftest`` imports ``nab.cli`` before any test runs, so
-        no test enters the package by another route.
+        The suite itself always enters through ``nab.cli``, since
+        ``conftest`` imports it before any test runs, so no test reaches
+        these modules in the order a caller of one of them would.
         """
         probe = subprocess.run(  # noqa: S603 - the probe is this file's own source
             [sys.executable, "-c", statement],
@@ -6543,7 +6621,9 @@ class TestDownloadCommand:
         out = tmp_path / "vendor"
         download_result = DownloadResult(written=(out / "x.whl",), skipped=())
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch(
                 "nab._download.download_lock", return_value=download_result
             ) as mock_dl,
@@ -6560,7 +6640,9 @@ class TestDownloadCommand:
         out = tmp_path / "vendor"
         download_result = DownloadResult(written=(), skipped=())
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch(
                 "nab._download.download_lock", return_value=download_result
             ) as mock_dl,
@@ -6579,7 +6661,7 @@ class TestDownloadCommand:
 
         with (
             patch(
-                "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
             ) as mock_resolve,
             patch(
                 "nab._download.download_lock", return_value=download_result
@@ -6608,7 +6690,7 @@ class TestDownloadCommand:
         transport = _ConcurrencyProbeTransport(payloads)
 
         with (
-            patch("nab._run.resolve_for_targets", return_value=result),
+            patch("nab._resolve.resolve_for_targets", return_value=result),
             patch("nab._download._make_transport", return_value=transport),
         ):
             download(pyproject, output=out, max_concurrency=cap)
@@ -6631,7 +6713,7 @@ class TestDownloadCommand:
         result, payloads = _fetchable_resolve_result(4)
         transport = _ConcurrencyProbeTransport(payloads)
         with (
-            patch("nab._run.resolve_for_targets", return_value=result),
+            patch("nab._resolve.resolve_for_targets", return_value=result),
             patch("nab._download._make_transport", return_value=transport),
         ):
             download(pyproject, output=out)
@@ -6649,7 +6731,7 @@ class TestDownloadCommand:
         def run() -> str:
             """Download into ``out`` and return what the run wrote to stderr."""
             with (
-                patch("nab._run.resolve_for_targets", return_value=result),
+                patch("nab._resolve.resolve_for_targets", return_value=result),
                 patch(
                     "nab._download._make_transport",
                     return_value=_ConcurrencyProbeTransport(payloads),
@@ -6734,7 +6816,9 @@ class TestDownloadCommand:
             lambda p: SourceRoots(project_dir=p.parent, pyproject=p),
         )
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch("nab._download.download_lock", return_value=download_result),
         ):
             download(pyproject, output=out, project_resolution="lowest")
@@ -6764,7 +6848,7 @@ class TestDownloadCommand:
         download_result = DownloadResult(written=(out / "foo.whl",), skipped=())
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_multi_tuple_universal_result(),
             ),
             patch(
@@ -6785,7 +6869,7 @@ class TestDownloadCommand:
         pyproject = _universal_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=ConfigError(
                     "exactly one of [extra 'cpu', extra 'gpu'] must be selected"
                 ),
@@ -6812,7 +6896,8 @@ class TestDownloadCommand:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets", side_effect=ResolutionError("conflict")
+                "nab._resolve.resolve_for_targets",
+                side_effect=ResolutionError("conflict"),
             ),
             pytest.raises(SystemExit, match="1"),
         ):
@@ -6826,7 +6911,7 @@ class TestDownloadCommand:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 side_effect=MissingExtraError(
                     "foo==1.0 does not provide extra 'nonexistent'"
                 ),
@@ -6843,7 +6928,9 @@ class TestDownloadCommand:
     ) -> None:
         pyproject = _make_pyproject(tmp_path)
         with (
-            patch("nab._run.resolve_for_targets", side_effect=KeyError("dependencies")),
+            patch(
+                "nab._resolve.resolve_for_targets", side_effect=KeyError("dependencies")
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             download(pyproject)
@@ -6855,7 +6942,8 @@ class TestDownloadCommand:
         pyproject = _make_pyproject(tmp_path)
         with (
             patch(
-                "nab._run.resolve_for_targets", side_effect=MissingHashError("no hash")
+                "nab._resolve.resolve_for_targets",
+                side_effect=MissingHashError("no hash"),
             ),
             pytest.raises(SystemExit, match="1"),
         ):
@@ -6867,7 +6955,9 @@ class TestDownloadCommand:
     ) -> None:
         pyproject = _make_pyproject(tmp_path)
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             patch(
                 "nab._download.download_lock", side_effect=DownloadError("sha mismatch")
             ),
@@ -6884,7 +6974,9 @@ class TestDownloadCommand:
         out = tmp_path / "wheels"
         out.write_text("not a directory")
         with (
-            patch("nab._run.resolve_for_targets", return_value=_stub_resolve_result()),
+            patch(
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
+            ),
             pytest.raises(SystemExit, match="1"),
         ):
             download(pyproject, output=out)
@@ -6952,7 +7044,7 @@ class TestDownloadCommand:
         download_result = DownloadResult(written=(), skipped=())
         with (
             patch(
-                "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
             ) as mock_resolve,
             patch("nab._download.download_lock", return_value=download_result),
         ):
@@ -6968,7 +7060,7 @@ class TestDownloadCommand:
         download_result = DownloadResult(written=(), skipped=())
         with (
             patch(
-                "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
             ) as mock_resolve,
             patch("nab._download.download_lock", return_value=download_result),
         ):
@@ -6984,7 +7076,7 @@ class TestDownloadCommand:
         download_result = DownloadResult(written=(), skipped=())
         with (
             patch(
-                "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
             ) as mock_resolve,
             patch("nab._download.download_lock", return_value=download_result),
         ):
@@ -7000,7 +7092,7 @@ class TestDownloadCommand:
         download_result = DownloadResult(written=(), skipped=())
         with (
             patch(
-                "nab._run.resolve_for_targets", return_value=_stub_resolve_result()
+                "nab._resolve.resolve_for_targets", return_value=_stub_resolve_result()
             ) as mock_resolve,
             patch("nab._download.download_lock", return_value=download_result),
         ):
@@ -7012,7 +7104,7 @@ class TestDownloadCommand:
         download_result = DownloadResult(written=(), skipped=())
         with (
             patch(
-                "nab._run.resolve_for_targets",
+                "nab._resolve.resolve_for_targets",
                 return_value=_multi_tuple_universal_result(),
             ) as mock_resolve,
             patch("nab._download.download_lock", return_value=download_result),
